Add BackUpRepository abstract class
